Mer varningar som är fixade!

+++ source/Pressure/pcgsolver/blas_win.h
@@ -6,7 +6,7 @@ namespace BLASWIN{
 	template <typename T>
 	T abs_max(std::vector<T> &v){
 		double max = fabs(v[0]);
-		for(int i=1; i<v.size(); ++i){
+		for(int i=1; i<(int)v.size(); ++i){
 			if(fabs(v[i]) > max)
 				max = fabs(v[i]);
 		}
@@ -16,7 +16,7 @@ namespace BLASWIN{
 	template <typename T>
 	T dot(std::vector<T> &v1, std::vector<T> &v2){
 		double sum=0;
-		for(int i=0; i< v1.size(); ++i){
+		for(int i=0; i< (int)v1.size(); ++i){
 			sum += v1[i]*v2[i];
 		}
 
@@ -25,7 +25,7 @@ namespace BLASWIN{
 
 	template <typename T>
 	void add_scaled(double alpha, std::vector<T> &s, std::vector<T> &res){
-		for(int i=0; i<s.size(); ++i){
+		for(int i=0; i<(int)s.size(); ++i){
 			res[i] = s[i] * alpha + res[i];
 		}
 		return;
